Move pat_and_expr_can_be_question_mark into clippy_utils
